[Flink] Extract shared lifecycle into AbstractFlinkJob to fix remaining duplication
